// ui/src/generated/ComponentAgentEvent.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ComponentAgentEventContent } from "./ComponentAgentEventContent";

export type ComponentAgentEvent = { run_id: string, sequence: number, created_at_ms: number, content: ComponentAgentEventContent, };